+++ tests/traits_impl_option_result/mod.rs
@@ -1,4 +1,3 @@
 mod applicative_laws;
-mod comonad_laws;
 mod functor_laws;
 mod monad_laws;